move popup hook from wrong directory
