+++ packages/client/src/hooks/index.ts
@@ -1,3 +1,4 @@
 export { default as useInput } from 'hooks/useInput'
 export { default as useAuth } from 'hooks/useAuth'
-export { default as useIntersectionObserver } from './useIntersectionObserver'
+export { default as useIntersectionObserver } from 'hooks/useIntersectionObserver'
+export { default as useOutsideClick } from 'hooks/useOutsideClick'
