+++ packages/client/src/hooks/useOutsideClick.ts
@@ -1,0 +1,17 @@
+import React, { useEffect } from 'react'
+
+const useOutsideClick = (ref: React.RefObject<HTMLElement>, callback: () => void) => {
+  const handleClick = (e: MouseEvent) => {
+    if (ref.current && e.target instanceof Node && !ref.current.contains(e.target)) {
+      callback()
+    }
+  }
+
+  useEffect(() => {
+    document.addEventListener('click', handleClick)
+    return () => {
+      document.removeEventListener('click', handleClick)
+    }
+  })
+}
+export default useOutsideClick
